move appdefs into pkg/applications
