+++ pkg/applications/defaultcatalog/applicationdefinitions/argocd-app.yaml
