+++ pkg/applications/defaultcatalog/applicationdefinitions/flux2-app.yaml
