+++ pkg/applications/defaultcatalog/applicationdefinitions/kube-vip-app.yaml
