+++ pkg/applications/defaultcatalog/applicationdefinitions/nginx-app.yaml
