+++ pkg/applications/defaultcatalog/applicationdefinitions/trivy-app.yaml
