+++ pkg/applications/defaultcatalog/applicationdefinitions/trivy-operator-app.yaml
